Add audio

+++ src/Random.h
@@ -1,6 +1,9 @@
 #ifndef RANDOM_H
 #define RANDOM_H
 
+#include <ctime>
+#include <random>
+
 namespace RandomConstants {
 	static const double inverseRandMaxD = 1. / static_cast <double> (RAND_MAX);
 	static const double inverseRandMaxF = 1.f / static_cast <float>  (RAND_MAX);
@@ -10,12 +13,12 @@ namespace RandomConstants {
 class Random {
 public:
 	// Initialize some static values
-	static void init() { std::srand(std::time(nullptr)); }
+	static void init() { std::srand(static_cast<unsigned int>(std::time(nullptr))); }
 	static void init(unsigned int seed) { std::srand(seed); }
 
 	static float randomFloat(float LO, float HI) {
 		if (LO == HI) return LO;
-		return LO + static_cast <float> (rand()) * (HI - LO) * RandomConstants::inverseRandMaxF;
+		return LO + static_cast <float> (rand()) * (HI - LO) * static_cast<float>(RandomConstants::inverseRandMaxF);
 	}
 
 	static double randomDouble(double LO, double HI) {
